// frontend/src/app/core/enums/instructor_tabs.ts
export enum InstructorTabs{
    DASHBOARD = '/instructor/instructor-dashboard',
    COURSE = '/instructor/course',
    PERFORMANCE = '/instructor/performance',
    NOTIFICATIONS = '/instructor/notifications',
    PAYMENT = '/instructor/payment',
    AFFILIATE = 'AFFILIATE',
    AFFILIATE_PROFILES = '/instructor/affiliate/profiles',
    PREMIUM_COURSES = '/instructor/affiliate/premium-courses',
    PREMIUM_STUDENTS="/instructor/premium-student",
    COURSE_CONTENT = "/content-type",
    AI_GRADER_CLASSES = "/ai-grader/classes",
    AI_GRADER_ASSESSMENTS = "/ai-grader/assessments",
    GRADER_RESULTS = "/grader-results",
    AI_GRADER = 'AI_GRADER'
}